pyproject.toml
[tool.poetry]
name = "repl_python3_statix"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
markdown = "^3.3.4"
toml = "^0.10.2"
libsass = "^0.20.1"
jinja2 = "^2.11.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
